savepoints support
This code appears to be unnecessary. SQLite is using the sub-journal to deal with the savepoints.
In the case of cache spill, the pages are written to the sub-journal and to the WAL/LMDB.
In the case of a rollback to a savepoint, the page is reloaded from the sub-journal and
the last version of the pages are written to the WAL before the commit.